Make retries more gentle and have heartbeat upload report success
With the upload returning a bool the nonlocal variable is no longer
reset, meaning you get free retries the next time around the loop,
hence making the retry extremely gentle.

Also, now that we're sure that the cache-control stuff works, remove
the reloading of the blobs and asserting that the cache control is
correct.