Update the MetadataServer post refactor
Updates the old MetadataServer to use a TimedMetadataServer for
uploading, and turns the existing class into a new one called
a MetadataCreator which looks at each new image and creates a shard
for it as it lands. Note the TimedMetadataServer also collects shards
from other processes which write to the same shard dir.

This architecture should allow this to easily mine the EFD
(or other summit DBs once they exist) for info, should that become
necessary, which is why this information scraping wasn't shoe-horned
into the isrRunner, where it would effectively have been able to run
for free (i.e. without having to open every image again). Once the
image info is *fully* available via a butler or other DBs, we can
remove that i/o for the sake of efficiency.