Add Sasquatch schema files for clean room LSSTCam, TS8, and LATISS
Schema is per instrument.

Required fields include:  "Observation reason", "Science program",
"Filter", and "Disperser".

Currently these fields are tags as configured in the Kafka->InfluxDB
connector for lsst.rubintv database on usdf-dev and hence required.
If they are missing, the kafka messages would not get written into
the InfluxDB database.

We can change what fields to use as the tags later. For now, just
add these fields in the schema.