Remove dataId files and FileWatchers entirely
This breaks botTesting.py but that's all old and outdated and
due for removal. Enough replacement code has been put in that
mypy doesn't complain about it, but if that was to be revived
it would need real pod definitions, but also so many other things
that it's not worth doing now. Should be scheduled for removal
in a few weeks, most probably.